Add drivers interfaces and core streaming datastructure
Here, we update drivers plugin with two interfaces:

* A high level `ExecTaskStreamingDriver` interface, to be implemented by basic
  drivers
* A low level API for streaming exec requests and responses, intended to be used
when a driver delegates nomad exec handling to a further processor.

Additionally, we establish the core data structure for the streaming operation
in the drivers protobuf (e.g. `ExecTaskStreamingRequest` and
`ExecTaskStreamingResponse`) and stream `ExecTaskStream` . These data structures
are to be used by drivers, clients, as well as API. We use the protobuf structs
directly to avoid translation and conversion overheads.